feat(endpoint): add password reset feature
- add columns to user table
- add endpoint to send reset email
- add endpoint to verify password reset link and new password
- write tests

[Delivers #164139683]
