--- .nyc_output/76c64d62-a9ed-41c2-80f1-279e09c5da1d.json
@@ -1,1 +1,0 @@
-{}
