+++ controllers/user.js
@@ -1,7 +1,12 @@
+import bcrypt from 'bcrypt';
 import dotenv from 'dotenv';
 import jwt from 'jsonwebtoken';
-import { compareSync } from 'bcrypt';
+import shortId from 'shortid';
+import sendMail from '../helpers/emails';
+import getName from '../helpers/user';
 import { User } from '../models';
+
+const { HOST_URL } = process.env;
 
 dotenv.config();
 const { JWT_SECRET } = process.env;
@@ -79,7 +84,7 @@ export default class UserController {
       body: { rememberMe, password },
       user
     } = req;
-    const passwordMatch = compareSync(password, user.password);
+    const passwordMatch = bcrypt.compareSync(password, user.password);
     if (!passwordMatch) {
       return res.status(401).json({
         success: false,
@@ -136,6 +141,104 @@ export default class UserController {
         success: false,
         message: 'User not found'
       });
+    });
+  }
+
+  /**
+   * @description reset user password
+   * @param {object} req http request object
+   * @param {object} res http response object
+   * @returns {object} response
+   */
+  static async resetPassword(req, res) {
+    const { email } = req.body;
+    const passwordResetToken = shortId.generate();
+    try {
+      // save password reset token to db
+      await User.update(
+        { passwordResetToken },
+        { where: { email } }
+      );
+
+      // save token expiration date (1 hr) to db
+      await User.update(
+        { passwordResetTokenExpires: Date.now() + (60 * 60 * 1000) },
+        { where: { email } }
+      );
+    } catch (error) {
+      return res.status(500).json({
+        success: false,
+        errors: [error.message]
+      });
+    }
+
+    // get user's name and send reset email
+    const name = await getName(email);
+    const emailPayload = {
+      name,
+      email,
+      link: `${HOST_URL}/api/v1/verifypasswordkey/${passwordResetToken}`,
+      subject: 'Reset your password',
+      message: 'reset your password'
+    };
+    sendMail(emailPayload);
+
+    return res.status(201).json({
+      success: true,
+      message: 'A link to reset your password has been sent to your mail. Please note that the link is only valid for one hour.'
+    });
+  }
+
+  /**
+   * @description Check if password token in valid
+   * @param {object} req http request object
+   * @param {object} res http response object
+   * @returns {object} response
+   */
+  static async checkPasswordToken(req, res) {
+    const { params: { passwordResetToken } } = req;
+    const { password } = req.body;
+    const getPasswordResetToken = await User.findOne({ where: { passwordResetToken } });
+    if (!getPasswordResetToken) {
+      return res.status(404).json({
+        success: false,
+        errors: ['Password reset token not found']
+      });
+    }
+    if (getPasswordResetToken.dataValues.passwordResetTokenExpires < Date.now()) {
+      return res.status(410).json({
+        success: false,
+        errors: ['Your link has expired. Please try to reset password again.']
+      });
+    }
+    try {
+      // save new password to db
+      await User.update(
+        { password: bcrypt.hashSync(password, bcrypt.genSaltSync(10)) },
+        { where: { passwordResetToken } }
+      );
+    } catch (error) {
+      return res.status(500).json({
+        success: false,
+        errors: [error.message]
+      });
+    }
+
+    try {
+      // delete password token
+      await User.update(
+        { passwordResetToken: '' },
+        { where: { passwordResetToken } }
+      );
+    } catch (error) {
+      return res.status(500).json({
+        success: false,
+        errors: [error.message]
+      });
+    }
+    return res.status(201).json({
+      success: true,
+      message: 'Password changed successfully.'
     });
   }
 }
